Handle migration of labels
Labels can now be specified on the components. These labels will be converted
into their respective component resource fields in theInstallation.
